# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Construction Ledger API
  description: >
    REST API for the Construction Ledger app. Projects come in three kinds —
    'site' (construction sites, credit/debit), 'account' (Srishti Construction /
    Sri Srishti Construction, withdrawal-only, reflects credits tagged to them
    from site payments), and 'loan' (credit = disbursement, debit = repayment).
  version: 3.0.0
servers:
  - url: http://localhost:4000
    description: Local Docker dev
  # - url: https://your-backend-url.example.com
  #   description: Production
...
paths:
...
  /api/transactions:
    get:
      tags: [Transactions]
      summary: List transactions, optionally filtered by project
      description: >
        For kind='account' projects, this returns a UNION of direct withdrawals
        plus credit entries reflected from sites via received_into_project_id —
        running_balance for accounts is credit-minus-withdrawal, computed
        server-side, not the raw window-function value.
      security: [{ bearerAuth: [] }]
      parameters:
        - name: project_id
          in: query
          required: false
          schema: { type: string, format: uuid }
          description: Filter to one project. Omit to list across everything.
      responses:
        '200':
          description: List of transactions with running_balance
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Transaction' }
    post:
      tags: [Transactions]
      summary: Create a transaction
      description: >
        credit always forces category="Client Payment" server-side regardless
        of what's sent. debit/withdrawal must use a category from
        GET /api/transactions/meta/categories, with subcategory required or
        forbidden depending on that category's own subcategory list.
        received_into_project_id is required when type=credit on a kind=site
        project, and must reference a kind=account project.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project_id, type, amount]
              properties:
                project_id: { type: string, format: uuid }
                date: { type: string, format: date-time, description: Defaults to now if omitted }
                type: { type: string, enum: [credit, debit, withdrawal] }
                category: { type: string, example: "Construction Material", description: Ignored/overridden when type=credit }
                subcategory: { type: string, example: Cement, nullable: true }
                detail: { type: string, example: "Rajesh Contractor" }
                amount: { type: number, example: 9920 }
                entered_by: { type: string, example: ishwar }
                received_into_project_id: { type: string, format: uuid, description: "Required when type=credit on a site project" }
      responses:
        '201':
          description: Transaction created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  txn_id: { type: string, format: uuid }
                  data: { $ref: '#/components/schemas/Transaction' }
        '400':
          description: Missing/invalid fields, invalid category+subcategory combo, or type not valid for this project's kind
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/transactions/{id}:
    patch:
      tags: [Transactions]
      summary: Update a transaction
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TxnId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                date: { type: string, format: date-time }
                type: { type: string, enum: [credit, debit, withdrawal] }
                category: { type: string }
                subcategory: { type: string, nullable: true }
                detail: { type: string }
                amount: { type: number }
                received_into_project_id: { type: string, format: uuid }
      responses:
        '200':
          description: Updated transaction
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { $ref: '#/components/schemas/Transaction' }
        '404':
          description: Transaction not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    delete:
      tags: [Transactions]
      summary: Delete a transaction
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/TxnId'
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SuccessResponse' }
        '404':
          description: Transaction not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /api/transactions/meta/categories:
    get:
      tags: [Transactions]
      summary: Get the full category → subcategory taxonomy
      description: Returns the nested EXPENSE_CATEGORIES structure, not a flat list. Credit entries don't use this — they're always fixed to "Client Payment".
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Category taxonomy
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        category: { type: string, example: Plumbing }
                        subcategories:
                          type: array
                          items: { type: string }
                          example: ['Labour', 'Material', 'Sanitary & Accessories']
  /api/transactions/{id}/period-summary:
    get:
      tags: [Transactions]
      summary: "⚠ Currently mounted under /api/transactions — see note"
      description: >
        NOTE: this route's own code comment says "GET /api/projects/:id/period-summary",
        but it's registered inside the transactions router, so the real,
        currently-working path is /api/transactions/:id/period-summary
        (documented here). Move it into projects.js to match the intended
        path and the rest of the project-summary endpoints. For kind=account:
        {total_credit (reflected, in range), total_withdrawal (in range),
        leftover = credit − withdrawal}. For other kinds: plain credit-debit
        for the range.
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: from
          in: query
          required: true
          schema: { type: string, format: date }
        - name: to
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        '200':
          description: Period summary
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: object
                    properties:
                      project_id: { type: string, format: uuid }
                      kind: { type: string }
                      from: { type: string, format: date }
                      to: { type: string, format: date }
                      total_credit: { type: number }
                      total_debit: { type: number }
                      total_withdrawal: { type: number }
                      leftover: { type: number }
        '400':
          description: Missing from/to query params
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Project not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Obtain via POST /api/auth/login.
  parameters:
    ProjectId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    TxnId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  schemas:
...
    Transaction:
      type: object
      properties:
        txn_id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        date: { type: string, format: date-time }
        type: { type: string, enum: [credit, debit, withdrawal] }
        category: { type: string, example: "Construction Material" }
        subcategory: { type: string, example: Cement, nullable: true }
        detail: { type: string, example: "Rajesh Contractor" }
        amount: { type: number, example: 9920 }
        entered_by: { type: string, example: ishwar }
        received_into_project_id: { type: string, format: uuid, nullable: true, description: "Set only on credit entries from a site — which account received the payment" }
        created_at: { type: string, format: date-time }
        running_balance: { type: number, example: 71600 }
    SuccessResponse:
      type: object
      properties:
        success: { type: boolean, example: true }
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        error: { type: string, example: "Descriptive error message" }
security:
  - bearerAuth: []